Improve the installation script

+++ install.sh
@@ -1,10 +1,5 @@
 #!/bin/bash
 
-if [ -f ~/.vimrc ]; then
-    echo ".vimrc exists." >&2
-    exit 1
-fi
-
 if ! hash brew 2>/dev/null; then
     echo "brew not found." >&2
     exit 1
@@ -18,12 +13,23 @@ if [ -z "$(brew ls --versions the_silver_searcher)"  ]; then
     brew install the_silver_searcher
 fi
 
+abspath() {
+    [[ $1 = /* ]] && echo "$1" || echo "$PWD/${1#./}"
+}
+
+if [ -f ~/.vimrc ]; then
+    if [ "$(readlink ~/.vimrc)" != "$(abspath ~/.vim/.vimrc)" ]; then
+        echo ".vimrc exists." >&2
+        exit 1
+    fi
+else
+    ln -s ~/.vim/.vimrc ~/.vimrc
+fi
+
 pushd ~/.vim > /dev/null
 git submodule init
 git submodule update
 popd > /dev/null
-
-ln -s ~/.vim/.vimrc ~/.vimrc
 
 vim +PluginInstall +qall
